+++ CHANGELOG.md
@@ -1,11 +1,21 @@
-## [0.2.8]
+# Changelog
+
+## 0.2.9
+- [x] Add examples
+- [x] Add imports
+- [x] Add `caller` field to the model input (can be: `run`, `autorun` or a button name)
+- [x] Add `title` field (for buttons rn)
+- [x] If `display` field is `false` the input is not shown
+- [x] If `autorun` is true, then actually autorun the model initially
+
+## 0.2.8
 - [x] Fix no input case
 
-## [0.2.7]
+## 0.2.7
 - [x] Show output when result is `0`
 - [x] Updated style for buttons and inputs
 
-## [<0.2.6]
+## 0.2.6
 - [x] Tests
 - [x] Load schema from query (loader)
 - [x] Reset button appears only after data change